requirements.pypy.txt
anyascii==0.2.0
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.2.4
django-appconf==1.0.4
django-cogwheels==0.3
django-comments-xtd==2.9.2
django-compressor==2.4.1
django-contrib-comments==2.1.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-libsass==0.8
django-modelcluster==5.1
django-taggit==1.4.0
django-treebeard==4.5.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docutils==0.17.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
greenlet==0.4.13
gunicorn==20.1.0
html5lib==1.1
idna==2.10
l18n==2020.6.1
libsass==0.21.0
openpyxl==3.0.7
Pillow==8.2.0
psutil==5.8.0
pytz==2021.1
rcssmin==1.0.6
readline==6.2.4.1
requests==2.25.1
rjsmin==1.1.0
sentry-sdk==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.1.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wagtail==2.13.2
wagtail-color-panel==1.2.0
wagtailmenus==3.0.2
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
